Final Code

+++ src/main/java/org/common/repository/CategoryRepository.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
-    Optional<Category> findById(Integer id);
+    Optional<Category> findById(Long id);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 
 }
